Adds a NOTICE file in the distributed application
To comply with redistribution licenses used in the app, we add a NOTICE file
that contains the dependencies (and transitive dependencies) of the libraries
that we are using in production.

This NOTICE does not include development dependencies. It's also pushed to
the repo to make easy for other people to check it.

Some things done in this commit:

* splits js dependencies between devDependencies and dependencies to allow
  the automatic tools track this
* Changes the scripts that bundled the app to automate the generation of
  the NOTICE
* Adds a link in the README to the repo and includes the README in the bundle to
  provide a way to access the source code
* Small script to only put in the NOTICE production dependencies and not all the
  virtualenv as happens with python tool that is being used